+++ packages/picasso/src/test-utils/index.tsx
@@ -1,18 +1,12 @@
-import React, { ReactElement } from 'react'
+import { ReactElement } from 'react'
 import { render, RenderOptions, RenderResult } from '@testing-library/react'
-import Picasso from '@toptal/picasso-shared'
+
+import TestingPicasso from './TestingPicasso'
 
 const customRender = (
   ui: ReactElement,
   options?: Omit<RenderOptions, 'queries'>
-): RenderResult => {
-  return render(
-    <Picasso loadFavicon={false} loadFonts={false} fixViewport={false}>
-      {ui}
-    </Picasso>,
-    options
-  )
-}
+): RenderResult => render(ui, { wrapper: TestingPicasso, ...options })
 
 export * from '@testing-library/react'
-export { customRender as render }
+export { customRender as render, TestingPicasso }